requirements.txt
customtkinter>=5.2.0
pyyaml>=6.0
paramiko>=3.0.0
psutil>=5.9.0
requests>=2.31.0
Pillow>=10.0.0
